refactor axon_map, scoreboard into beyeler2019
