+++ pulse2percept/models/setup.py
@@ -12,12 +12,8 @@ def configuration(parent_package='', top_path=None):
         libraries.append('m')
 
     if platform.python_implementation() != 'PyPy':
-        config.add_extension('_scoreboard',
-                             sources=['_scoreboard.pyx'],
-                             include_dirs=[numpy.get_include()],
-                             libraries=libraries)
-        config.add_extension('_axon_map',
-                             sources=['_axon_map.pyx'],
+        config.add_extension('_beyeler2019',
+                             sources=['_beyeler2019.pyx'],
                              include_dirs=[numpy.get_include()],
                              libraries=libraries)
         config.add_extension('_nanduri2012',
